requirements.txt
pandas==1.2.2
flask==1.1.2
psycopg2-binary==2.8.6
SQLAlchemy==1.3.23